(WIP) Beginning of task view table
I need to take some time to think through the logic of how to create
empty <td> elements for months where a given employee has no assignment
to this task.

Also need to consider whether it would be possible to have an assignment
start/end date that is prior to or after the start/end date of the SOW
associated with the given task. If yes, lots more logic needed to
generate the correct <th> headers.