Fix bug in re-fetching of taskViewData
Async re-fetch of taskViewData was firing based off of a local boolean.
By the time the async returns, the boolean was always true, resulting in
many more re-fetches then necessary. Abstracting the assignment submital
to a function and passing the boolean as an argument results in the
desired behavior of only a single re-fetch of the data